requirements.txt

Flask==2.2.5
google-cloud-vision==3.4.5
